Add 16x16 squares into the grid with JS
The height and width of the squares are calculated based on the
height of the grid divided by 16 minus the border width of the grid.
The border width of the grid is divided with 16 so each square gets a
proportional cut that  makes it possible to fit 16 squares on a line.